Move DroppedVariableStats to CodeGen lib
To get Dropped variable statistics for MIR, we need to move the base
class DroppedVariableStats code to the CodeGen library because we
cannot have CodeGen link against Passes.
